pyproject.toml
[tool.poetry]
name = "PhoenixBios"
version = "0.1.0"
description = ""
authors = ["taimaskhanov11 <81189712+taimaskhanov11@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
PyYAML = "^6.0"
PyAutoGUI = "^0.9.53"
loguru = "^0.6.0"
Pillow = "^9.0.1"
keyboard = "^0.13.5"
pywinauto = "^0.6.8"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
auto-py-to-exe = "^2.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
